requirements.txt
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
wikipedia
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain-huggingface
faiss-cpu
langchain-chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
langchain
streamlit
langchain-streamlit 